Chrome: Extract some code to windows_utils.py
